Move IsContainerObject from API to SPI
fixes #248
