+++ docker/docker/src/main/java/org/arquillian/cube/docker/impl/client/containerobject/CubeContainerObjectTestEnricher.java
@@ -19,7 +19,6 @@ import org.arquillian.cube.containerobject.Cube;
 import org.arquillian.cube.containerobject.CubeDockerFile;
 import org.arquillian.cube.containerobject.HostPort;
 import org.arquillian.cube.containerobject.Image;
-import org.arquillian.cube.containerobject.IsContainerObject;
 import org.arquillian.cube.containerobject.Link;
 import org.arquillian.cube.docker.impl.client.config.BuildImage;
 import org.arquillian.cube.docker.impl.client.config.CubeContainer;
@@ -31,6 +30,7 @@ import org.arquillian.cube.docker.impl.util.DockerFileUtil;
 import org.arquillian.cube.impl.util.ReflectionUtil;
 import org.arquillian.cube.spi.Binding;
 import org.arquillian.cube.spi.CubeRegistry;
+import org.arquillian.cube.spi.metadata.IsContainerObject;
 import org.jboss.arquillian.core.api.Injector;
 import org.jboss.arquillian.core.api.Instance;
 import org.jboss.arquillian.core.api.annotation.Inject;
